move load counters
