+++ corehq/apps/change_feed/data_sources.py
@@ -20,7 +20,7 @@ from corehq.form_processor.document_stores import (
 )
 from corehq.util.couch import get_db_by_doc_type
 from corehq.util.couchdb_management import couch_config
-from corehq.util.datadog.utils import (
+from corehq.util.metrics.load_counters import (
     case_load_counter,
     form_load_counter,
     ledger_load_counter,
