+++ corehq/messaging/scheduling/scheduling_partitioned/dbaccessors.py
@@ -6,7 +6,7 @@ from corehq.sql_db.util import (
     get_db_aliases_for_partitioned_query,
     paginate_query_across_partitioned_databases,
 )
-from corehq.util.datadog.utils import load_counter_for_model
+from corehq.util.metrics.load_counters import load_counter_for_model
 
 
 def _validate_class(obj, cls):
@@ -171,7 +171,7 @@ def _paginate_query_across_partitioned_databases(model_class, q_expression, load
 
 
 def _paginate_query(db_name, model_class, q_expression, load_source, query_size=5000):
-    track_load = load_counter_for_model(model_class)(load_source, None, extra_tags=['db:{}'.format(db_name)])
+    track_load = load_counter_for_model(model_class)(load_source, None, extra_tags={'db': db_name})
     sort_cols = ('active', 'next_event_due')
 
     # active is always set to true in the queryset's q_expression so we
